Revert "Port System.Diagnostics.PerformanceCounters to .NET Core"
