+++ src/Common/src/Interop/Windows/kernel32/Interop.WaitForSingleObject.cs
@@ -6,11 +6,11 @@ using System;
 using System.Runtime.InteropServices;
 using Microsoft.Win32.SafeHandles;
 
-internal partial class Interop
+internal static partial class Interop
 {
-    internal partial class Kernel32
+    internal static partial class Kernel32
     {
-        [DllImport(Libraries.Kernel32, ExactSpelling=true, SetLastError=true)]
-        internal static extern int WaitForSingleObject(SafeWaitHandle handle, int timeout);
+        [DllImport(Interop.Libraries.Kernel32, ExactSpelling = true, SetLastError = true, EntryPoint = "WaitForSingleObject")]
+        public static extern int WaitForSingleObject(SafeWaitHandle handle, int timeout);
     }
 }
