--- src/Common/src/Microsoft/Win32/SafeHandles/SafeLocalMemHandle.cs
@@ -1,23 +1,0 @@
-// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-// See the LICENSE file in the project root for more information.
-
-using System;
-
-namespace Microsoft.Win32.SafeHandles
-{
-    internal sealed class SafeLocalMemHandle : SafeHandleZeroOrMinusOneIsInvalid
-    {
-        internal SafeLocalMemHandle() : base(true) { }
-
-        internal SafeLocalMemHandle(IntPtr existingHandle, bool ownsHandle) : base(ownsHandle)
-        {
-            SetHandle(existingHandle);
-        }
-
-        protected override bool ReleaseHandle()
-        {
-            return Interop.Kernel32.LocalFree(handle) == IntPtr.Zero;
-        }
-    }
-}
