+++ src/System.Diagnostics.PerformanceCounter/dir.props
@@ -2,7 +2,6 @@
 <Project ToolsVersion="14.0" DefaultTargets="Build" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
   <Import Project="..\dir.props" />
   <PropertyGroup>
-    <AssemblyVersion>4.0.0.0</AssemblyVersion>
-    <AssemblyKey>Open</AssemblyKey>
+    <AssemblyVersion>4.0.1.0</AssemblyVersion>
   </PropertyGroup>
 </Project>
