--- src/System.Diagnostics.PerformanceCounter/src/Configurations.props
@@ -1,10 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<Project ToolsVersion="14.0" DefaultTargets="Build" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
-  <PropertyGroup>
-    <BuildConfigurations>
-      netcoreapp-Windows_NT;
-      netstandard;
-      netfx;
-    </BuildConfigurations>
-  </PropertyGroup>
-</Project>
