--- src/System.Diagnostics.PerformanceCounter/src/System/Diagnostics/PerformanceCounterInstanceLifetime.cs
@@ -1,13 +1,0 @@
-// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-// See the LICENSE file in the project root for more information.
-
-namespace System.Diagnostics
-{
-    public enum PerformanceCounterInstanceLifetime
-    {
-        Global = 0,
-        Process = 1
-    }
-}
-
