--- src/System.Diagnostics.PerformanceCounter/tests/Configurations.props
@@ -1,9 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<Project ToolsVersion="14.0" DefaultTargets="Build" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
-  <PropertyGroup>
-    <BuildConfigurations>
-      netcoreapp-Windows_NT;
-      netfx;
-    </BuildConfigurations>
-  </PropertyGroup>
-</Project>
