+++ src/System.Diagnostics.Process/src/System/Diagnostics/Process.Windows.cs
@@ -107,37 +107,6 @@ namespace System.Diagnostics
             // Nop
         }
 
-        /// <devdoc>
-        ///     Make sure we are watching for a process exit.
-        /// </devdoc>
-        /// <internalonly/>
-        private void EnsureWatchingForExit()
-        {
-            if (!_watchingForExit)
-            {
-                lock (this)
-                {
-                    if (!_watchingForExit)
-                    {
-                        Debug.Assert(_haveProcessHandle, "Process.EnsureWatchingForExit called with no process handle");
-                        Debug.Assert(Associated, "Process.EnsureWatchingForExit called with no associated process");
-                        _watchingForExit = true;
-                        try
-                        {
-                            _waitHandle = new Interop.Kernel32.ProcessWaitHandle(_processHandle);
-                            _registeredWaitHandle = ThreadPool.RegisterWaitForSingleObject(_waitHandle,
-                                new WaitOrTimerCallback(CompletionCallback), null, -1, true);
-                        }
-                        catch
-                        {
-                            _watchingForExit = false;
-                            throw;
-                        }
-                    }
-                }
-            }
-        }
-
         /// <summary>
         /// Instructs the Process component to wait the specified number of milliseconds for the associated process to exit.
         /// </summary>
@@ -150,7 +119,7 @@ namespace System.Diagnostics
                 if (handle.IsInvalid)
                     return true;
 
-                using (Interop.Kernel32.ProcessWaitHandle processWaitHandle = new Interop.Kernel32.ProcessWaitHandle(handle))
+                using (ProcessWaitHandle processWaitHandle = new ProcessWaitHandle(handle))
                 {
                     return _signaled = processWaitHandle.WaitOne(milliseconds);
                 }
@@ -215,7 +184,7 @@ namespace System.Diagnostics
                         // since some process could return an actual STILL_ACTIVE exit code (259).
                         if (!_signaled) // if we just came from WaitForExit, don't repeat
                         {
-                            using (var wh = new Interop.Kernel32.ProcessWaitHandle(handle))
+                            using (var wh = new ProcessWaitHandle(handle))
                             {
                                 _signaled = wh.WaitOne(0);
                             }
@@ -796,7 +765,7 @@ namespace System.Diagnostics
                     // Since haveProcessHandle is true, we know we have the process handle
                     // open with at least SYNCHRONIZE access, so we can wait on it with 
                     // zero timeout to see if the process has exited.
-                    using (Interop.Kernel32.ProcessWaitHandle waitHandle = new Interop.Kernel32.ProcessWaitHandle(_processHandle))
+                    using (ProcessWaitHandle waitHandle = new ProcessWaitHandle(_processHandle))
                     {
                         if (waitHandle.WaitOne(0))
                         {
